Store Compression: integer overflow causes failed reads (index is safe), closes #2071.

+++ src/main/java/org/elasticsearch/common/compress/CompressedIndexInput.java
@@ -187,7 +187,7 @@ public abstract class CompressedIndexInput extends IndexInput {
             return false;
         }
         currentOffsetIdx++;
-        currentUncompressedChunkPointer = currentOffsetIdx * uncompressed.length;
+        currentUncompressedChunkPointer = ((long) currentOffsetIdx) * uncompressed.length;
         position = 0;
         return (position < valid);
     }
